feat(ts): stage 7 — convert element components to typescript
Nine Two.js-bound React components. Loose ElementProps + ShapeLike `any`
types throughout — these props come from ElementRenderWrapper / canvas /
factories whose internals stay loose during the migration. Stage 12 cleanup
will sweep what we can after newCanvas converges in Stage 9.

Converted:
- src/components/elements/circle.tsx
- src/components/elements/rectangle.tsx
- src/components/elements/diamond.tsx
- src/components/elements/pencil.tsx
- src/components/elements/arrowLine.tsx
- src/components/elements/divider.tsx
- src/components/elements/newText.tsx (largest — text input overlay,
  proportional resize via corner handles)
- src/components/elements/groupobject.tsx — focused-group orchestration;
  Two.js scene.subtractions safety pattern from CLAUDE.md preserved
- src/components/utils/elementRenderWrappers.tsx — typed as
  ComponentType<ElementProps> factories

Common patterns introduced:
- `useImmer<Record<string, any>>({})` for the internalState bags
- DOM getElementById results captured into `const el = ...; if (el) ...`
  blocks rather than chained `.style` access; satisfies strict null checks
- Event handlers receive concrete DOM types (FocusEvent, KeyboardEvent,
  MouseEvent) at the boundary; everything past `event.target` falls back to
  `any` via cast.
- Two.js constructor-namespace access (Two.Anchor, Two.Commands) goes
  through `(Two as any)` since @types/two.js doesn't expose those.
- factoryModules glob updated to '.ts' to match Stage 3's converted
  factories.

yarn typecheck passes.
