--- src/components/elements/arrowLine.js
@@ -1,333 +1,0 @@
-import React, { useEffect, useState, useRef } from 'react'
-
-import PropTypes from 'prop-types'
-import { useImmer } from 'use-immer'
-import Two from 'two.js'
-
-import ElementCreator from '../../factory/arrowLine'
-
-import { updateX1Y1Vertices, updateX2Y2Vertices } from '../../utils/updateVertices'
-import { strokeTypeToDashes } from '../../utils/misc'
-import { useMediaQueryUtils } from '../../constants/exportHooks'
-
-function ArrowLine(props) {
-    const [showToolbar, toggleToolbar] = useState(false)
-    const [internalState, setInternalState] = useImmer({})
-    const stateRefForGroup = useRef()
-    const { isMobile } = useMediaQueryUtils()
-
-    const two = props.twoJSInstance
-    let selectorInstance = null
-    let groupObject = null
-
-    function onBlurHandler(e) {
-        if (
-            e?.relatedTarget?.id === 'floating-toolbar' ||
-            e?.relatedTarget?.dataset.parent === 'floating-toolbar'
-        ) {
-            const getGroupElementFromDOM = document.getElementById(
-                `${stateRefForGroup.current.id}`
-            )
-            // set the focus and on blur recursively until no floating toolbar touch is observed
-            getGroupElementFromDOM.focus()
-            getGroupElementFromDOM.addEventListener('blur', onBlurHandler)
-        } else {
-            selectorInstance.pointCircle1Group.opacity = 0
-            selectorInstance.pointCircle2Group.opacity = 0
-            two.update()
-            document.getElementById(`${groupObject.id}`) &&
-                document
-                    .getElementById(`${groupObject.id}`)
-                    .removeEventListener('keydown', handleKeyDown)
-        }
-    }
-
-    function handleKeyDown(e) {
-        if (e.keyCode === 8 || e.keyCode === 46) {
-            // DELETE/BACKSPACE KEY WAS PRESSED
-            document.getElementById(`${groupObject.id}`)?.blur()
-            props.handleDeleteComponent &&
-                props.handleDeleteComponent(groupObject)
-            two.remove([groupObject])
-            two.update()
-        }
-    }
-
-    function onFocusHandler(e) {
-        document.getElementById(`${groupObject.id}`).style.outline = 0
-        document
-            .getElementById(`${groupObject.id}`)
-            .addEventListener('keydown', handleKeyDown)
-    }
-
-    // Using unmount phase to remove event listeners
-    useEffect(() => {
-        // Calculate x and y through dividing width and height by 2 or vice versa
-        // if x and y are given then multiply width and height into 2
-        let hitAreaObserver = null
-        const offsetHeight = 0
-        const prevX = props.x
-        const prevY = props.y
-
-        // console.log('arrowLine props', props)
-        // Instantiate factory
-        const elementFactory = new ElementCreator(two, prevX, prevY, {
-            x1: props.x1 ?? 20,
-            x2: props.x2 ?? 100,
-            y1: props.y1 ?? 10,
-            y2: props.y2 ?? 10,
-            strokeType: props.strokeType,
-            linewidth: props.linewidth,
-            isMobile,
-        })
-        // Get all instances of every sub child element
-        const {
-            group,
-            pointCircle1Group,
-            pointCircle2Group,
-            pointCircle1,
-            pointCircle2,
-            line,
-        } = elementFactory.createElement()
-        group.elementData = { ...props.itemData, ...props }
-        line.opacity = props.metadata?.opacity ?? 1
-        if (props.stroke) line.stroke = props.stroke
-        if (props.linewidth) line.linewidth = props.linewidth
-
-        if (props.parentGroup) {
-            /** This element will be rendered and scoped in its parent group */
-            const parentGroup = props.parentGroup
-            parentGroup.add(group)
-            two.update()
-        } else {
-            /** This element will render by creating it's own group wrapper */
-            groupObject = group
-            stateRefForGroup.current = group
-            // resizeLineInstance = resizeLine
-            // const { selector } = getEditComponents(two, group, 4)
-            selectorInstance = {
-                pointCircle1Group,
-                pointCircle2Group,
-            }
-
-            two.update()
-
-            // Wider transparent hit area — inset from each endpoint so the
-            // tail/head circles remain independently clickable
-            const lineDomElem = document.getElementById(line.id)
-            if (lineDomElem) {
-                const HIT_INSET = 20
-                const hitElem = document.createElementNS('http://www.w3.org/2000/svg', 'line')
-                hitElem.setAttribute('stroke', 'transparent')
-                hitElem.setAttribute('stroke-width', '12')
-                hitElem.setAttribute('pointer-events', 'stroke')
-                hitElem.style.cursor = 'pointer'
-
-                function syncHitLine() {
-                    const v0 = line.vertices[0]
-                    const v1 = line.vertices[1]
-                    const dx = v1.x - v0.x
-                    const dy = v1.y - v0.y
-                    const len = Math.sqrt(dx * dx + dy * dy)
-                    if (len < HIT_INSET * 2 + 5) return
-                    const r = HIT_INSET / len
-                    hitElem.setAttribute('x1', v0.x + dx * r)
-                    hitElem.setAttribute('y1', v0.y + dy * r)
-                    hitElem.setAttribute('x2', v1.x - dx * r)
-                    hitElem.setAttribute('y2', v1.y - dy * r)
-                }
-
-                syncHitLine()
-                lineDomElem.parentNode.insertBefore(hitElem, lineDomElem)
-                hitAreaObserver = new MutationObserver(syncHitLine)
-                hitAreaObserver.observe(lineDomElem, { attributes: true, attributeFilter: ['d'] })
-            }
-
-            // some styling
-            document.getElementById(line.id).style.cursor = 'pointer'
-            document.getElementById(pointCircle1Group.id).style.cursor =
-                'pointer'
-            document.getElementById(pointCircle2Group.id).style.cursor =
-                'pointer'
-
-            // document
-            //     .getElementById(pointCircle2.id)
-            //     .setAttribute('class', 'dragger-picker')
-
-            document
-                .getElementById(group.id)
-                .setAttribute('class', 'dragger-picker')
-
-            // 1. for point circle 1
-            document
-                .getElementById(pointCircle1Group.id)
-                .setAttribute('class', 'dragger-picker is-line-circle')
-
-            document
-                .getElementById(pointCircle1Group.id)
-                .setAttribute('data-parent-id', group.id)
-
-            document
-                .getElementById(pointCircle1Group.id)
-                .setAttribute('data-line-id', line.id)
-
-            document
-                .getElementById(pointCircle1Group.id)
-                .setAttribute('data-direction', 'left')
-
-            // 2. for point circle 2
-            document
-                .getElementById(pointCircle2Group.id)
-                .setAttribute('class', 'dragger-picker is-line-circle')
-
-            document
-                .getElementById(pointCircle2Group.id)
-                .setAttribute('data-parent-id', group.id)
-
-            document
-                .getElementById(pointCircle2Group.id)
-                .setAttribute('data-line-id', line.id)
-
-            document
-                .getElementById(pointCircle2Group.id)
-                .setAttribute('data-direction', 'right')
-
-            document
-                .getElementById(group.id)
-                .setAttribute('data-component-id', props.id)
-            document
-                .getElementById(group.id)
-                .setAttribute('data-linewidth', String(props.linewidth ?? ''))
-
-            setInternalState((draft) => {
-                draft.element = {
-                    [group.id]: group,
-                    [line.id]: line,
-                }
-                draft.group = {
-                    id: group.id,
-                    data: group,
-                }
-                draft.shape = {
-                    id: line.id,
-                    data: line,
-                }
-                draft.line = {
-                    id: line.id,
-                    data: line,
-                }
-                draft.pointCircle1Group = {
-                    id: pointCircle1Group.id,
-                    data: pointCircle1Group,
-                }
-                draft.pointCircle2Group = {
-                    id: pointCircle2Group.id,
-                    data: pointCircle2Group,
-                }
-                draft.text = {
-                    data: {},
-                }
-                draft.icon = {
-                    data: {},
-                }
-            })
-
-            const getGroupElementFromDOM = document.getElementById(
-                `${group.id}`
-            )
-            getGroupElementFromDOM.addEventListener('focus', onFocusHandler)
-            getGroupElementFromDOM.addEventListener('blur', onBlurHandler)
-        }
-
-        return () => {
-            hitAreaObserver?.disconnect()
-            two.remove(group)
-        }
-    }, [])
-
-    // Syncs visual properties to the Two.js shape when props change.
-    // Props are updated by ElementRenderWrapper (newCanvas.js) whenever componentStore changes,
-    // which happens on GraphQL subscription updates from other users editing this component.
-    useEffect(() => {
-        if (internalState?.group?.data) {
-            let groupInstance = internalState.group.data
-            let lineInstance = internalState.line.data
-            let pointCircle1Group = internalState.pointCircle1Group.data
-            let pointCircle2Group = internalState.pointCircle2Group.data
-            groupInstance.translation.x = props.x
-            groupInstance.translation.y = props.y
-            lineInstance.opacity = props.metadata?.opacity ?? 1
-            two.update()
-            // updateX1Y1Vertices(lineInstance, props.x1, props.y1, pointCircle1Group)
-            // updateX2Y2Vertices(lineInstance, props.x2, props.y2, pointCircle2Group)
-        }
-    }, [props.x, props.y, props.metadata])
-
-    // hook for same sync behavior discussed in comment as above
-    useEffect(() => {
-        if (internalState?.line?.data) {
-            let lineInstance = internalState.line.data
-            if (props.stroke) lineInstance.stroke = props.stroke
-            if (props.linewidth) lineInstance.linewidth = props.linewidth
-            lineInstance.dashes = strokeTypeToDashes(props.strokeType)
-            two.update()
-        }
-    }, [props.stroke, props.linewidth, props.strokeType])
-
-    // hook for same sync behavior discussed in comment as above
-    useEffect(() => {
-        if (internalState?.group?.data) {
-            updateX1Y1Vertices(
-                Two,
-                internalState.line.data,
-                props.x1,
-                props.y1,
-                internalState.pointCircle1Group.data,
-                two
-            )
-            updateX2Y2Vertices(
-                Two,
-                internalState.line.data,
-                props.x2,
-                props.y2,
-                internalState.pointCircle2Group.data,
-                two
-            )
-        }
-    }, [props.x1, props.x2, props.y1, props.y2])
-
-    function closeToolbar() {
-        toggleToolbar(false)
-    }
-
-    return (
-        <React.Fragment>
-            {/* {showToolbar ? (
-                <Toolbar
-                    toggle={showToolbar}
-                    componentState={internalState}
-                    closeToolbar={closeToolbar}
-                    componentId={props.id}
-                    postToolbarUpdate={() => {
-                        two.update()
-                    }}
-                    // updateComponent={() => {
-                    //     two.update()
-                    // }}
-                />
-            ) : null} */}
-        </React.Fragment>
-    )
-}
-
-ArrowLine.propTypes = {
-    x: PropTypes.number,
-    y: PropTypes.number,
-}
-
-// ArrowLine.defaultProps = {
-//     x: 100,
-//     y: 50,
-// }
-
-export default ArrowLine
